[14.0][IMP] contract: Add test to check next_period_date_start on former lines / split tests
