+++ contract/tests/__init__.py
@@ -1,4 +1,5 @@
 from . import test_contract
 from . import test_contract_manually_create_invoice
 from . import test_portal
+from . import test_stop_contract
 from . import test_multicompany
